move common Preview related code to a :common module
