+++ common/build.gradle.kts
@@ -1,0 +1,29 @@
+plugins {
+    kotlin("jvm") version "2.0.0"
+    id("maven-publish")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+}
+
+dependencies {
+    api(project(":core"))
+    implementation(libs.classgraph)
+    implementation(libs.kotlin.reflect)
+}
+
+//https://www.talentica.com/blogs/publish-your-android-library-on-jitpack-for-better-reachability/
+publishing {
+    publications {
+        create<MavenPublication>("mavenJava") {
+            afterEvaluate {
+                from(components["kotlin"])
+            }
+            groupId = "sergio.sastre.composable.preview.scanner"
+            artifactId = "common"
+            version = "0.6.1"
+        }
+    }
+}
