+++ settings.gradle.kts
@@ -28,5 +28,6 @@ include(":tests")
 include(":core")
 include(":android")
 include(":jvm")
+include(":common")
 include(":custompreviews")
 include(":paparazzi-plugin-tests")
