+++ tests/custom-paparazzi-tests.gradle.kts
@@ -5,6 +5,7 @@ tasks.register("paparazziPreviews") {
     description = "Records/verifies screenshots with Paparazzi for ComposablePreview tests"
     group = "Verification"
 
+    // For the sourceTests
     tasks.findByName("compileDebugScreenshotTestKotlin")?.let { dependsOn(it) }
     tasks.findByName("compileDebugAndroidTestKotlin")?.let { dependsOn(it) }
